+++ apps/daffio/src/app/docs/guides/components/guides-content/guides-content.component.html
@@ -1,3 +1,3 @@
 <daffio-doc-viewer [breadcrumbs]="doc().breadcrumbs" [toc]="doc().tableOfContents" [sourcePath]="doc().sourcePath">
-	<div daffDocsCopyCodeBlock [innerHTML]="doc().contents | safe"></div>
+	<div [innerHTML]="doc().contents | safe"></div>
 </daffio-doc-viewer>
