+++ libs/docs/src/copy-button/public_api.ts
@@ -1,2 +1,1 @@
-export { DaffDocsCopyButtonComponent } from './copy-button/copy-button.component';
-export { DaffDocsCopyCodeBlockDirective } from './copy-code-block/copy-code-block.directive';
+export { DaffDocsCopyButtonComponent } from './copy-button.component';
